First basic implementation: copy template file and open in Jupyter
A first basic implementation of `nbkickoff` that copies the template
file, and either opens the new notebook in an existing Jupyter server
process, or launches a new one.

Logic for detecting a running server and opening the notebook is loosely
based on `nbopen`, but with various tweaks and modifications. Most
importantly, `nbkickoff` will *always* terminate, even when launching a
new Jupyter server process. In contrast, `nbopen` *becomes* the Jupyter
server process if no suitable running server was found, which is
troublesome for our application.